Refined the apis.

+++ Sources/PlusAuth/Storage.swift
@@ -12,17 +12,17 @@ protocol Storage {
 	func updateUserVerification(userId: String, verified: Bool) throws
 
 	// Refresh Tokens
-	func insertRefreshToken(userId: String, token: String, expiration: Int) throws
-	func selectRefreshToken(token: String) throws -> (expiration: Int, userId: String)
+	func insertRefreshToken(userId: String, token: String, expiration: Int64) throws
+	func selectRefreshToken(token: String) throws -> (expiration: Int64, userId: String)
 	func deleteRefreshToken(token: String) throws
 
 	// Verify Account Tokens
-	func insertVerifyAccountToken(token: String, expiration: Int, userId: String) throws
-	func selectVerifyAccountToken(token: String) throws -> (expiration: Int, userId: String)
+	func insertVerifyAccountToken(token: String, expiration: Int64, userId: String) throws
+	func selectVerifyAccountToken(token: String) throws -> (expiration: Int64, userId: String)
 	func deleteVerifyAccountToken(token: String) throws
 
 	// Reset Password Tokens
-	func insertResetPasswordToken(token: String, expiration: Int, userId: String) throws
-	func selectResetPasswordToken(token: String) throws -> (expiration: Int, userId: String)
+	func insertResetPasswordToken(token: String, expiration: Int64, userId: String) throws
+	func selectResetPasswordToken(token: String) throws -> (expiration: Int64, userId: String)
 	func deleteResetPasswordToken(token: String) throws
 }
